UI: Fix slack chat hooks for Okta and Azure SSO (#6204)
