+++ openmetadata-ui/src/main/resources/ui/src/components/SlackChat/SlackChat.css
@@ -1,0 +1,3 @@
+.slack-chat > div > div > div {
+  z-index: 100;
+}
